fix: switch to float32 to eliminate bfloat16 NaN in optimizer step
Root cause: bfloat16 RMSNorm backward + AdamW on bfloat16 params produced NaN
gradients on the very first step. With CPUMasterModel only keeping 2 layers on
GPU at a time, float32 memory (~1.5GB peak) is still well under 12GB.

- LlamaConfig torch_dtype: bfloat16 -> float32
- AutoModelForCausalLM.from_config dtype: bfloat16 -> float32
- CPUMasterConfig dtype: bfloat16 -> float32

Reproduced locally: 2-layer test with bfloat16 -> 1 NaN param after first opt step.
Same test with float32 -> 0 NaN params.

+++ pretrain_megatrain.py
@@ -165,13 +165,13 @@ def main():
         attention_bias=False,
         mlp_bias=False,
         initializer_range=0.02,
-        torch_dtype="bfloat16",
+        torch_dtype="float32",
         head_dim=128,
         architectures=["LlamaForCausalLM"],
     )
     hf_model = AutoModelForCausalLM.from_config(
         hf_config,
-        dtype=torch.bfloat16,
+        dtype=torch.float32,
         trust_remote_code=True,
         attn_implementation="sdpa",
     )
@@ -215,7 +215,7 @@ def main():
         checkpoint_interval=args.checkpoint_interval,
         num_grad_slabs=args.num_grad_slabs,
         device=args.device,
-        dtype=torch.bfloat16,
+        dtype=torch.float32,
         log_interval=1,
         attn_implementation="sdpa",
         trust_remote_code=True,
